Add custom hook to prep related components data, toggle list items with switch too
