+++ moped-editor/src/views/projects/projectView/ProjectComponents/RelatedComponentListItem.js
@@ -47,6 +47,7 @@ export default function RelatedComponentListItem({
 
   const componentName = component?.moped_components?.component_name;
   const listItemPrimaryText = componentName;
+  console.log(component);
 
   return (
     <Box
